+++ app/Http/Controllers/api/v1/ApplicationController.php
@@ -57,6 +57,8 @@ class ApplicationController extends Controller
         setting()->set('room_limit', $request->room_limit);
         setting()->set('own_rooms_pagination_page_size', $request->own_rooms_pagination_page_size);
         setting()->set('pagination_page_size', $request->pagination_page_size);
+        setting()->set('password_self_reset_enabled', $request->password_self_reset_enabled);
+        setting()->set('default_timezone', $request->default_timezone);
         setting()->set('banner', array_filter($request->banner, function ($setting) {
             return $setting !== null;
         }));