+++ app/Http/Kernel.php
@@ -4,6 +4,7 @@ namespace App\Http;
 
 use App\Http\Middleware\EnsureModelNotStale;
 use App\Http\Middleware\RoomAuthenticate;
+use App\Http\Middleware\RouteEnableIf;
 use Illuminate\Foundation\Http\Kernel as HttpKernel;
 use Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful;
 
@@ -73,5 +74,6 @@ class Kernel extends HttpKernel
         'room.authenticate'     => RoomAuthenticate::class,
         'loggedin'              => \App\Http\Middleware\LoggedInUser::class,
         'check.stale'           => EnsureModelNotStale::class,
+        'enable_if'             => RouteEnableIf::class,
     ];
 }