+++ resources/js/components/Login/EmailLoginComponent.vue
@@ -27,7 +27,14 @@
           required
           :placeholder="passwordLabel"
           :state="errors !== null && errors.password && errors.password.length > 0 ? false: null"
+          aria-describedby="passwordHelpBlock"
         ></b-form-input>
+
+        <b-form-text id="passwordHelpBlock">
+          <router-link to="/forgot_password" v-if="settings('password_self_reset_enabled')">
+            {{ $t('auth.forgotPassword') }}
+          </router-link>
+        </b-form-text>
 
         <b-form-invalid-feedback v-if="errors !== null && errors.password && errors.password.length > 0">
           <template v-for="error in errors.password">
@@ -45,6 +52,8 @@
 </template>
 
 <script>
+import { mapGetters } from 'vuex';
+
 export default {
   props: [
     'errors',
@@ -55,6 +64,11 @@ export default {
     'title',
     'emailLabel'
   ],
+  computed: {
+    ...mapGetters({
+      settings: 'session/settings'
+    })
+  },
   data () {
     return {
       email: '',