+++ resources/lang/de/validation.php
@@ -246,5 +246,8 @@ return [
         'server_pool'                       => 'Serverpool',
         'icon_text'                         => 'Icon Text',
         'color'                             => 'Farbe',
+        'password_self_reset_enabled'       => 'Registrierten Nutzern ermöglichen das Passwort zurückzusetzen',
+        'timezone'                          => 'Zeitzone',
+        'default_timezone'                  => 'Standardzeitzone'
     ],
 ];