+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Artisan command to create a new admin user ([#81])
 - Application banner that can be modified through the application settings ([#89], [#94])
 - Management of server pools, used by load balancer and added to room type ([#96], [#99])
+- Password self reset, password reset by an administrator and password generation with password reset for registered users ([#106], [#107])
 - Logging for failed/successful logins and ldap roles ([#112], [#113])
 
 [#1]: https://github.com/THM-Health/PILOS/issues/1
@@ -86,6 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#102]: https://github.com/THM-Health/PILOS/pull/102
 [#103]: https://github.com/THM-Health/PILOS/issues/103
 [#105]: https://github.com/THM-Health/PILOS/pull/105
+[#106]: https://github.com/THM-Health/PILOS/issues/106
+[#107]: https://github.com/THM-Health/PILOS/pull/107
 [#112]: https://github.com/THM-Health/PILOS/issues/112
 [#113]: https://github.com/THM-Health/PILOS/pull/113